+++ src/main/kotlin/expresions/UMinusExpression.kt
@@ -2,7 +2,7 @@ package expresions
 
 import machine.ProgramMemory
 
-internal class UMinusExpression(var value: Int) : Expression {
+internal class UMinusExpression(var value: Long) : Expression {
     override fun toOpcodes() {
         ProgramMemory.push("PUSH $value")
         ProgramMemory.push("NEG")
